+++ tests/fixes/mod.rs
@@ -9,6 +9,7 @@ mod incorrect_delimiter;
 mod leading_character;
 mod space_character;
 mod trailing_whitespace;
+mod unordered_key;
 
 #[test]
 fn correct_file() {
@@ -52,97 +53,6 @@ fn key_without_value() {
 
     testdir.close();
 }
-
-#[test]
-fn unordered_key() {
-    let testdir = TestDir::new();
-    let testfile = testdir.create_testfile(
-        ".env",
-        "# comment\n\
-        \n\
-        C=B\n\
-        # A comment\n\
-        A=B\n\
-        X=Y\n\
-        # more comments\n\
-        \n\
-        # middle comment\n\
-        \n\
-        M=123\n\
-        # K comment\n\
-        K=123\n\
-        # I comment\n\
-        # I comment\n\
-        I=123\n\
-        \n\
-        A1=1\n\
-        B1=1\n\
-        \n\
-        # end comment\n",
-    );
-    let expected_output = String::from(
-        "Fixed warnings:\n\
-        .env:5 UnorderedKey: The A key should go before the C key\n\
-        .env:13 UnorderedKey: The K key should go before the M key\n\
-        .env:16 UnorderedKey: The I key should go before the K key\n",
-    );
-    testdir.test_command_fix_success(expected_output);
-
-    assert_eq!(
-        testfile.contents().as_str(),
-        "# comment\n\
-        \n\
-        # A comment\n\
-        A=B\n\
-        C=B\n\
-        X=Y\n\
-        # more comments\n\
-        \n\
-        # middle comment\n\
-        \n\
-        # I comment\n\
-        # I comment\n\
-        I=123\n\
-        # K comment\n\
-        K=123\n\
-        M=123\n\
-        \n\
-        A1=1\n\
-        B1=1\n\
-        \n\
-        # end comment\n",
-    );
-
-    testdir.close();
-}
-
-// We need to figure out how to test unfixed warnings now when all the fixers have been implemented
-// (if required)
-
-// #[test]
-// fn unfixed_warnings() {
-//     let testdir = TestDir::new();
-//     let testfile = testdir.create_testfile(".env", "A=DEF\nB=BAR \nX-Y=Z\nf=BAR\nA=FOO\n");
-//
-//     let expected_output = String::from(
-//         "Fixed warnings:\n\
-//         .env:2 TrailingWhitespace: Trailing whitespace detected\n\
-//         .env:3 IncorrectDelimiter: The X-Y key has incorrect delimiter\n\
-//         .env:4 LowercaseKey: The f key should be in uppercase\n\
-//         .env:5 DuplicatedKey: The A key is duplicated\n\
-//         \n\
-//         Unfixed warnings:\n\
-//        .env:5 UnorderedKey: The A key should go before the A key\n",
-//     );
-//     testdir.test_command_fix_fail(expected_output);
-//
-//     assert_eq!(
-//         testfile.contents().as_str(),
-//         "A=DEF\nB=BAR\nX_Y=Z\nF=BAR\n# A=FOO\n"
-//     );
-//
-//     testdir.close();
-// }
 
 #[test]
 fn multiple_files() {
